Renamed PiCAM to arducam in preparation for architectural changes
