+++ ArduCAM4Pi/arducam_ov5642_digital_camera.c
@@ -13,7 +13,7 @@
 #include <stdint.h>
 #include <time.h>
 #include "utft_spi.h"
-#include "picam.h"
+#include "arducam.h"
 
 #define BOOL int
 #define TRUE 1
