+++ ArduCAM4Pi/mt9d111_regs.h
@@ -1,6 +1,6 @@
 #ifndef MT9D111_REGS_H
 #define MT9D111_REGS_H
-#include "PiCAM.h"
+#include "arducam.h"
 
 /*******************************************************************************************
  #  Display resolution standards #
