+++ ArduCAM4Pi/ov7660_regs.h
@@ -1,6 +1,6 @@
 #ifndef OV7660_REGS_H
 #define OV7660_REGS_H
-#include "picam.h"
+#include "arducam.h"
 const struct sensor_reg OV7660_QVGA[] PROGMEM =
 {
 	{0x11, 0x83},
