+++ ArduCAM4Pi/ov7725_regs.h
@@ -1,6 +1,6 @@
 #ifndef OV7725_REGS_H
 #define OV7725_REGS_H
-#include "picam.h"
+#include "arducam.h"
 
 const struct sensor_reg OV7725_QVGA[] PROGMEM =
 {
